Moved `View` to `actors.rs`
